Fix leading zeros
